% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.load.R
\docType{package}
\name{install.load-package}
\alias{install.load}
\alias{install.load-package}
\title{install.load: Check, Install and Load CRAN Packages}
\description{
install.load provides the function `install_load` which checks the local R
library(ies) to see if the required package(s) is/are installed or not. If
the package(s) is/are not installed, then the package(s) will be installed
along with the required dependency(ies). This function pulls source or
binary packages from the Posit/RStudio-sponsored CRAN mirror Repository.
Lastly, the chosen package(s) is/are loaded. The function
`load_package` simply loads the provided package(s).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/install.load}
  \item Report bugs at \url{https://gitlab.com/iembry/install.load/-/issues}
}

}
\keyword{internal}
