% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/install_load.R
\name{install_lo}
\alias{install_lo}
\alias{install_load}
\title{Check, Install and Load CRAN & USGS GRAN Packages}
\source{
Check if R package is installed then load library answered by maloneypatr is the source for the original function. See \url{http://stackoverflow.com/questions/15155814/check-if-r-package-is-installed-then-load-library}.
}
\usage{
install_load(package1, ...)
}
\arguments{
\item{package1}{R package.}

\item{...}{Any additional R packages.}
}
\value{
Install and load \code{package1} and other packages listed in the arguments ... and all of the dependencies.
}
\description{
Checks the local R library(ies) to see if the required package(s) is/are installed or not. If the package(s) is/are not installed, then the package(s) will be installed along with the required dependency(ies). This function pulls source or binary packages from the Revolution Analytics CRAN mirror and/or the USGS GRAN Repository. Lastly, the chosen package(s) is/are loaded.
}
\examples{
\dontrun{
library(install.load)
install_load("chron")
install_load("chron", "data.table", "dataRetrieval", "EGRET", "dplyr")
}
}
\keyword{install}
\keyword{load}
\keyword{packages}

