% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package.R
\encoding{UTF-8}
\name{load_package}
\alias{load_package}
\title{Load Packages}
\source{
\enumerate{
   \item Check if R package is installed then load library answered by
maloneypatr is the source for the original function. See
\url{https://stackoverflow.com/questions/15155814/check-if-r-package-is-installed-then-load-library}.
   \item Error Handling in R By Nina Zumel, October 9, 2012, R-Bloggers. See
   \url{http://www.r-bloggers.com/error-handling-in-r/}.
}
}
\usage{
load_package(package1, ...)
}
\arguments{
\item{package1}{R package.}

\item{...}{Any additional R packages.}
}
\value{
Load \code{package1} and other packages listed in the arguments \code{...}.
}
\description{
Loads the listed packages only
}
\examples{
\dontrun{
library("install.load")
# This will only load the package(s)
load_package("chron")

load_package("chron", "data.table", "dataRetrieval", "EGRET", "dplyr")
}


}
\seealso{
\code{\link{install_load}} if you are wanting to install and load packages
}
\author{
maloneypatr, Irucka Embry
}
