% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_package.R
\name{load_package}
\alias{load_package}
\title{Load Packages}
\source{
Check if R package is installed then load library answered by
maloneypatr is the source for the original function. See
\url{http://stackoverflow.com/questions/15155814/check-if-r-package-is-installed-then-load-library}.
}
\usage{
load_package(package1, ...)
}
\arguments{
\item{package1}{R package.}

\item{...}{Any additional R packages.}
}
\value{
Load \code{package1} and other packages listed in the arguments \code{...}.
}
\description{
Loads packages [It is assumed that the package(s) that you are wishing to
load already exist in your library(ies). If not, then use the install_load
function to install and then load the packages.]
}
\examples{
\dontrun{
library(install.load)
# This will only load the package(s), so the packages should already be
# available in your library(ies)
load_package("chron")
load_package("chron", "data.table", "dataRetrieval", "EGRET", "dplyr")
}



}
\author{
maloneypatr, Irucka Embry
}
\seealso{
\code{\link{install_load}}
}

